+++ src/main/java/com/archers_expansion/items/armors/DeadeyeArmor.java
@@ -1,0 +1,36 @@
+package com.archers_expansion.items.armors;
+
+import com.archers_expansion.client.armor.DeadeyeRenderer;
+import com.archers_expansion.client.armor.TundraArcherRenderer;
+import mod.azure.azurelibarmor.animatable.GeoItem;
+import mod.azure.azurelibarmor.animatable.client.RenderProvider;
+import net.minecraft.client.render.entity.model.BipedEntityModel;
+import net.minecraft.entity.EquipmentSlot;
+import net.minecraft.entity.LivingEntity;
+import net.minecraft.item.ItemStack;
+import net.spell_engine.api.item.armor.Armor;
+import org.jetbrains.annotations.NotNull;
+
+import java.util.function.Consumer;
+
+public class DeadeyeArmor extends ModArmorItem implements GeoItem {
+    public DeadeyeArmor(Armor.CustomMaterial material, Type type, Settings settings) {
+        super(material, type, settings);
+    }
+
+    @Override
+    public void createRenderer(Consumer<Object> consumer) {
+        consumer.accept(new RenderProvider() {
+            private DeadeyeRenderer renderer;
+
+            @Override
+            public @NotNull BipedEntityModel<LivingEntity> getHumanoidArmorModel(LivingEntity livingEntity, ItemStack itemStack, EquipmentSlot equipmentSlot, BipedEntityModel<LivingEntity> original) {
+                if (renderer == null)
+                    renderer = new DeadeyeRenderer();
+
+                renderer.prepForRender(livingEntity, itemStack, equipmentSlot, original);
+                return renderer;
+            }
+        });
+    }
+}