+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+0.1.1
+- changed some damage multipliers for Enchanted Crystal Arrow and Arctic Volley
+- implemented new Poison Smoke Particles for Choking Gas
+- nerfed movement speed & ranged haste gain per stack for Fast Shot
+- buffed some spell power multipliers for War Archer Spells
+- added a Spell School, FROST_RANGE, so the Spell-Damage of the Tundra Archer Spell Book are better manageable
+- added Tundra Hunter Armor, gives Ranged Damage & Frost Spell Power
+- added Deadeye Armor, gives Ranged Damage, Draw Speed & Movement Speed
+- added War Archer Armor, which gives Ranged Damage and more Armor
+- added advancements
+- if the target is bleeding, Choking Poison deals more damage per tick
+
 0.1.0
 - First Beta Release of the Mod
 - the Mod only contains Spell Books and Spells, no equipment yet!